+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+* Ensure that modifier penalties and modifiers are always positive
+
 ## [0.0.7] - 2023-02-17
 
 ### Fixed